feat: Move ListChoiceScreen to kmpuiviews and Update Dependencies
This commit moves the `ListChoiceScreen` to the `kmpuiviews` module for multiplatform use, removes the old screen from `UIViews`, and updates related dependencies.

-   **`ListChoiceScreen`:**
    -   Moved `ListChoiceScreen` from `UIViews` to `kmpuiviews`.
    -   Removed the old `ListChoiceScreen` from `UIViews`.
+   - Updated `OptionsSheet` to use the new `ListChoiceScreen` from `kmpuiviews`.
+   - Updated `DetailsUtils` to use the new `ListChoiceScreen` from `kmpuiviews`.
+   - Updated `DeleteFromListScreen` to use `pluralStringResource`.
+   - General cleanup.
