+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/details/DetailsUtils.kt
@@ -75,10 +75,10 @@ import com.programmersbox.favoritesdatabase.ListDao
 import com.programmersbox.favoritesdatabase.NotificationItem
 import com.programmersbox.kmpmodels.KmpInfoModel
 import com.programmersbox.kmpuiviews.presentation.Screen
+import com.programmersbox.kmpuiviews.presentation.settings.lists.addtolist.ListChoiceScreen
 import com.programmersbox.uiviews.GenericInfo
 import com.programmersbox.uiviews.R
 import com.programmersbox.uiviews.presentation.components.ToolTipWrapper
-import com.programmersbox.uiviews.presentation.lists.ListChoiceScreen
 import com.programmersbox.uiviews.utils.launchCatching
 import com.programmersbox.uiviews.utils.showErrorToast
 import kotlinx.coroutines.CoroutineScope
