+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/lists/DeleteFromListScreen.kt
@@ -34,7 +34,7 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -57,7 +57,6 @@ fun DeleteFromListScreen(
     val logoDrawable = koinInject<AppLogo>()
     val navController = LocalNavController.current
     val dao = LocalCustomListDao.current
-    val context = LocalContext.current
     val scope = rememberCoroutineScope()
     val sheetState = rememberModalBottomSheetState()
     val onDismiss: () -> Unit = {
@@ -88,7 +87,7 @@ fun DeleteFromListScreen(
                 title = { Text("Delete") },
                 text = {
                     Text(
-                        context.resources.getQuantityString(
+                        pluralStringResource(
                             R.plurals.areYouSureRemove,
                             itemsToDelete.size,
                             itemsToDelete.size
